Moving validation checks into validation_checks.py from data_cleaning
