+++ mbs_results/validation_checks.py
@@ -18,6 +18,8 @@ def validate_config(config):
             """Overlapping column names in responses_keep_cols and and
             contributors_keep_cols (main config)."""
         )
+    validate_config_datatype_input(**config)
+    validate_config_repeated_datatypes(**config)
 
 
 def colnames_clash(
@@ -77,4 +79,74 @@ def validate_indices(responses, contributors):
         raise ValueError(
             f"""There are mismatched records between the responses and
             contributors datasets:\n {non_matches}"""
+        )
+
+
+def validate_config_datatype_input(
+    responses_keep_cols, contributors_keep_cols, **config
+):
+    """
+    function to validate config datatypes inputted into the config file
+
+    Parameters
+    ----------
+    responses_keep_cols : dict
+        dictionary containing columns to keep from responses and datatypes
+    contributors_keep_cols : _type_
+        dictionary containing columns to keep from contributors and datatypes
+
+    Raises
+    ------
+    ValueError
+        ValueError if the specified datatype is not in the accepted_type list
+        ints and floats do not need to specify number of bits i.e. int32 etc.
+    """
+    joint_dict = {**responses_keep_cols, **contributors_keep_cols}
+    accepted_types = ["str", "float", "int", "date", "bool", "category"]
+    incorrect_datatype = [
+        x for x in list(joint_dict) if joint_dict.get(x) not in accepted_types
+    ]
+
+    if incorrect_datatype:
+        given_types = [joint_dict.get(key) for key in incorrect_datatype]
+        raise ValueError(
+            "Check the inputted datatype(s) for column(s) {}:{},\
+            only the following datatypes are accepted: {}".format(
+                incorrect_datatype, given_types, accepted_types
+            )
+        )
+
+
+def validate_config_repeated_datatypes(
+    responses_keep_cols, contributors_keep_cols, **config
+):
+    """
+    Checking that repeated columns do not have conflicting data types
+
+    Parameters
+    ----------
+    responses_keep_cols : dict
+        dictionary containing columns to keep from responses and datatypes
+    contributors_keep_cols : _type_
+        dictionary containing columns to keep from contributors and datatypes
+
+    Raises
+    ------
+    ValueError
+        ValueError if any repeated column has a different data type in the contributors
+        and responses data sets.
+    """
+
+    mismatched_types = [
+        x
+        for x in responses_keep_cols
+        if (x in contributors_keep_cols)
+        and (responses_keep_cols[x] != contributors_keep_cols[x])
+    ]
+    if mismatched_types:
+        # Warning to catch if the same column name has different types
+        raise ValueError(
+            "Mismatched data types between two dictionaries in columns: {}".format(
+                mismatched_types
+            )
         )
